% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{buildModel}
\alias{buildModel}
\title{dddModel.buildModel}
\usage{
buildModel(path, Timeresinsec)
}
\arguments{
\item{path}{path of the directory where to find the two files}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}
}
\description{
Build the models used in the model ddd
Two files are required:
- one for the calibrated parameters (paramCalibrated.R)
- one for the GIS-related parameters (paramGIS.R)
The files are present in the data folder of this package.
}
\examples{
\dontrun{
buildModel()
}
}
\keyword{model}

