% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModel.R
\name{getModel}
\alias{getModel}
\title{Get the models to run ddd}
\usage{
getModel(method = NULL, path = NULL, inputParam = NULL, SAVE = FALSE,
  pathResults = "~/")
}
\arguments{
\item{method}{Method to get the model parameters: "buildNVE" or "load"}

\item{path}{Directory where to get the files}

\item{inputParam}{List input parameters}

\item{SAVE}{Save the results, Boolean}

\item{pathResults}{Path of the results. By default: $HOME}
}
\value{
a list of all the models used to run ddd
}
\description{
Get all the models to run ddd
Two options:
- build from a set of parameters
- load from a rda file
}
\examples{
\dontrun{
getModel()
}
}
\keyword{model}

