% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputParam.R
\docType{data}
\name{inputParam}
\alias{inputParam}
\title{List of input parameters from daily discharge at Narsjo}
\format{A rda file}
\description{
List of parameters used to build the ddd models
Values must be taken as illustrative
}
\keyword{datasets}

