% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{models}
\alias{models}
\title{List of models from daily discharge at Narsjo}
\format{A rda file}
\description{
List of models used to run the ddd models
Values must be taken as illustrative
}
\keyword{datasets}

