% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramCalibrated.R
\docType{data}
\name{paramCalibrated}
\alias{paramCalibrated}
\title{Set of calibrated parameters from daily discharge at Narsjo}
\format{A R file}
\source{
\href{http://www.nve.no}{NVE}
}
\description{
Set of calibrated parameters from daily discharge at Narsjo.
Some parameters might be erroneous.
The set of parameters is not to be taken as robust but as an example
}
\references{
the Norwegian Water Resources and Energy Directorate (NVE)
}
\keyword{datasets}

