% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramGIS.R
\docType{data}
\name{paramGIS}
\alias{paramGIS}
\title{Set of GIS parameters related to the catchement at Narsjo}
\format{A R file}
\source{
\href{http://www.nve.no}{NVE}
}
\description{
Set of GIS parameters related to the catchement at Narsjo
Some parameters might be erroneous.
The set of parameters is not to be taken as robust but as an example
}
\references{
the Norwegian Water Resources and Energy Directorate (NVE)
}
\keyword{datasets}

