% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramNVE.R
\docType{data}
\name{paramNVE}
\alias{paramNVE}
\title{Set of parameters from daily discharge at Narsjo}
\format{Both csv and tab files are avalaible.}
\source{
\href{http://www.nve.no}{NVE}
}
\description{
Set of parameters from daily discharge at Narsjo.
Values must be taken as illustrative
The set of parameters is not to be taken as robust but as an example
This file can be used as a template
}
\references{
the Norwegian Water Resources and Energy Directorate (NVE)
}
\keyword{datasets}

