% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModel.R
\name{getModel}
\alias{getModel}
\title{Get the models to run ddd}
\usage{
getModel(getModel, path, Timeresinsec = NULL)
}
\arguments{
\item{getModel}{method to get the model parameters ("build", "load" or "source")}

\item{path}{directory where to get the files}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)
The value can be NULL if the getModel method is "load" or "source"}
}
\value{
a list of all the models used to run ddd
}
\description{
Get all the models to run ddd
Three options:
- build from a set of parameters
- load from a rda file
- load from a R-file
}
\examples{
\dontrun{
getModel()
}
}
\keyword{model}

