% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParam.R
\name{getParam}
\alias{getParam}
\title{Get and process the parameters}
\usage{
getParam(Timeresinsec, path)
}
\arguments{
\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{path}{path of the directory where to find the two files}
}
\value{
The output is a list of processed parameters to be build the model
}
\description{
Get and process the parameters from a R-source file.
Two files are required:
- one for the calibrated parameters (paramCalibrated.R)
- one for the GIS-related parameters (paramGIS.R)
The files are present in the data folder of this package.
}
\examples{
\dontrun{
getParam()
}
}
\keyword{model}

